backend routes refactored
