+++ backend-flask/lib/cloudwatch.py
@@ -12,9 +12,10 @@ from flask import request
 # LOGGER.info('logger is working')
 
 def init_cloudwatch(response):
-  # CloudWatch LOGS
-  # @app.after_request
-  # def after_request(response):
   timestamp = strftime('[%Y-%b-%d %H:%M]')
   LOGGER.error('%s %s %s %s %s %s', timestamp, request.remote_addr, request.method, request.scheme, request.full_path, response.status)
   return response
+
+  # @app.after_request
+  # def after_request(response):
+  #   init_cloudwatch(response)
